+++ framework/server/src/test/java/is/codion/framework/server/EntityServerTest.java
@@ -26,7 +26,7 @@ import is.codion.common.rmi.server.ClientLog;
 import is.codion.common.rmi.server.RemoteClient;
 import is.codion.common.rmi.server.Server;
 import is.codion.common.rmi.server.ServerConfiguration;
-import is.codion.common.rmi.server.WhitelistObjectInputFilterFactory;
+import is.codion.common.rmi.server.WhitelistInputFilterFactory;
 import is.codion.common.rmi.server.exception.ConnectionNotAvailableException;
 import is.codion.common.rmi.server.exception.LoginException;
 import is.codion.common.rmi.server.exception.ServerAuthenticationException;
@@ -330,7 +330,7 @@ public class EntityServerTest {
 		ServerConfiguration.RMI_SERVER_HOSTNAME.set("localhost");
 		ServerConfiguration.KEYSTORE.set("src/main/config/keystore.jks");
 		ServerConfiguration.KEYSTORE_PASSWORD.set("crappypass");
-		WhitelistObjectInputFilterFactory.SERIALIZATION_FILTER_WHITELIST.set("classpath:serialization-whitelist-test.txt");
+		WhitelistInputFilterFactory.SERIALIZATION_FILTER_WHITELIST.set("classpath:serialization-whitelist-test.txt");
 
 		return EntityServerConfiguration.builder(3223, 3221)
 						.adminPort(3223)
@@ -341,7 +341,7 @@ public class EntityServerTest {
 						.domainClassNames(asList("is.codion.framework.server.TestDomain", "is.codion.framework.server.ConfigureDb"))
 						.clientLogging(true)
 						.sslEnabled(true)
-						.objectInputFilterFactoryClassName("is.codion.common.rmi.server.WhitelistObjectInputFilterFactory")
+						.objectInputFilterFactoryClassName(WhitelistInputFilterFactory.class.getName())
 						.build();
 	}